+++ src/whatsapp/dto/sendMessage.dto.ts
@@ -61,6 +61,7 @@ export class SendPollDto extends Metadata {
 export type MediaType = 'image' | 'document' | 'video' | 'audio';
 export class MediaMessage {
   mediatype: MediaType;
+  mimetype?: string;
   caption?: string;
   // for document
   fileName?: string;